Fix the pip3 settings in the FreeBSD yml files
The current versions of FreeBSD do not use pip-3.8 anymore. Let's
switch to the unversioned "pip" command here which should simply
point to the current version (i.e. pip-3.11 on FreeBSD 14.1).

While we're at it, update the image_name of FreeBSD 14 to use
the 14.1 instead of the 14.0 release, since that's the current one.

Signed-off-by: Thomas Huth <thuth@redhat.com>