cirrus-run: Avoid hard job failure when CI credits run out
Recently the libvirt project started to run out of CI credits on cirrus
towards the end of the month. As this is annoying we'd prefer if the job
is allowed to fail in such case.

This patch (temporarily) switches to 'cirrus-run' version with patches
allowing detection of CI credit failure, which also returns a different
error code allowing us to allow return code '3' to be considered an
allowed failure.

Signed-off-by: Peter Krempa <pkrempa@redhat.com>