lcitool: Fix cirrus job formatting for 'out of credits' handling
Commit d6d3834440b0fe8105cd3b572b9c011672b17ac6 added handling for
out of credists situation on 'Cirrus CI' by attempting to handle the
return code '3' from 'cirrus-run' as 'allow-failure'. Unforunately the
definition in the 'template' section was wrong as it gets overwritten by
the individual job definitions which pass an explicit 'allow-failure'
statement which completely overwrites the value from the template.

To handle this properly we need to do either:

 allow_failure: true

or in case when w don't want to disallow failure except when the script
returns '3':

 allow_failure:
   exit_codes: 3

Signed-off-by: Peter Krempa <pkrempa@redhat.com>